+++ client/src/app/services/dto/game.dto.ts
@@ -1,7 +1,9 @@
 import { Moment } from 'moment';
+import { UserProjection } from "./user.dto";
 
 export interface GameProjection {
     id?: number;
     name: string;
     startDateTime?: Moment;
+    players: UserProjection[];
 }
