+++ client/src/app/services/dto/group.dto.ts
@@ -1,4 +1,3 @@
-import { Moment } from 'moment';
 import { UserProjection } from './user.dto';
 import { GameProjection } from './game.dto';
