+++ server/src/main/resources/messages.properties
@@ -12,5 +12,3 @@ email.invalid = "Provided email is not correct."
 
 updatePassword.blank.newPassword = New password must not be blank.
 updatePassword.blank.oldPassword = Old password must not be blank.
-
-registration.blank.first
